Midando o tipo dos daods para decimal

+++ Calculadora/Operacoes.cs
@@ -3,9 +3,10 @@ namespace Calculadora
 {
     public class Operacoes
     {
-        public int valorA {get;set;}
+        // Ajuste dos tipos de dados para decimal
+        public decimal valorA { get; set; }
         public char operador { get; set; }
-        public int valorB { get; set; }
+        public decimal valorB { get; set; }
         public decimal resultado { get; set; }
 
     }
